fix signed/unsigned warning

+++ ssl/ssl_lib.c
@@ -1075,7 +1075,7 @@ long SSL_ctrl(SSL *s,int cmd,long larg,void *parg)
 		s->max_cert_list=larg;
 		return(l);
 	case SSL_CTRL_SET_MTU:
-		if (larg < dtls1_min_mtu())
+		if (larg < (long)dtls1_min_mtu())
 			return 0;
 
 		if (SSL_version(s) == DTLS1_VERSION ||
